refactor: provider and the context are separated
