+++ src/app/providers/ProductsProvider/ProductsProvider.tsx
@@ -1,18 +1,13 @@
+import { ProductsContext } from './ProductsContext'
 import type { JSX } from 'react'
-import { createContext, type ReactNode, useState } from 'react'
+import { type ReactNode, useState } from 'react'
 import { productsState, type ProductsStateType } from './ProductsState'
 
-export type IProductsContext = {
-  products: ProductsStateType
-  setProducts: (newProducts: ProductsStateType) => void
+type ProductsProviderProperties = {
+  children: ReactNode
 }
 
-export const ProductsContext = createContext<IProductsContext>({
-  products: productsState,
-  setProducts: (): void => {},
-})
-
-export function ProductsProvider({ children }: { children: ReactNode }): JSX.Element {
+export function ProductsProvider({ children }: ProductsProviderProperties): JSX.Element {
   const [products, setProductsState] = useState<ProductsStateType>(productsState)
 
   const setProducts = (newProducts: ProductsStateType): void => {
